pyproject.toml
[tool.poetry]
name = "analysis-project-generator"
version = "0.1.0"
description = ""
authors = ["Nils Strelow <strelow.nils@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<3.10"
fastapi = {extras = ["all"], version = "^0.63.0"}
copier = {git = "https://github.com/copier-org/copier.git"}
python-dotenv = "^0.17.0"
python-gitlab = "^2.6.0"
supabase-py = {git = "https://github.com/nstrelow/supabase-py.git"}

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
